unrelated: move max width toggle into its own component...
everything else in the header is split out into its own component, so we should probably do the same for the max width button.
