+++ resources/js/components/global-header/Header.vue
@@ -1,17 +1,12 @@
 <script setup>
-import { inject } from 'vue';
-import { Icon } from '@ui';
 import Logo from './Logo.vue';
 import Breadcrumbs from './Breadcrumbs.vue';
 import SiteSelector from './SiteSelector.vue';
 import Search from './Search.vue';
+import SupportButton from "@/components/global-header/SupportButton.vue";
+import MaxWidthButton from "@/components/global-header/MaxWidthButton.vue";
 import ViewSiteButton from './ViewSiteButton.vue';
 import UserDropdown from './UserDropdown.vue';
-import SupportButton from "@/components/global-header/SupportButton.vue";
-
-const layout = inject('layout', {});
-const isMaxWidthEnabled = layout.isMaxWidthEnabled;
-const toggleMaxWidth = layout.toggleMaxWidth;
 </script>
 
 <template>
@@ -35,15 +30,7 @@ const toggleMaxWidth = layout.toggleMaxWidth;
                 <Search />
             </div>
 	        <SupportButton />
-            <button
-                @click="toggleMaxWidth"
-                :aria-label="isMaxWidthEnabled ? __('Expand Layout') : __('Constrain Layout')"
-                v-tooltip="isMaxWidthEnabled ? __('Expand Layout') : __('Constrain Layout')"
-                class="hidden [@media(min-width:1800px)]:inline-flex items-center justify-center whitespace-nowrap shrink-0 font-medium antialiased cursor-pointer no-underline disabled:text-white/60 dark:disabled:text-white/50 disabled:cursor-not-allowed [&_svg]:shrink-0 [&_svg]:text-gray-925 [&_svg]:opacity-60 dark:[&_svg]:text-white bg-transparent hover:bg-gray-400/10 text-gray-900 dark:text-gray-300 dark:hover:bg-white/15 dark:hover:text-gray-200 h-8 text-[0.8125rem] leading-tight rounded-lg px-0 gap-0 w-8 [&_svg]:size-4 -me-2 [&_svg]:text-white/85! will-change-transform"
-                data-expand-layout-control
-            >
-                <Icon :name="isMaxWidthEnabled ? 'zoom-fit-screen' : 'fit-screen'" class="animate-pulse-on-appearance" />
-            </button>
+	        <MaxWidthButton />
             <ViewSiteButton />
             <UserDropdown />
         </div>
